rds-ad/requirements.txt
uvicorn==0.32.1
fastapi==0.115.5
boto3==1.35.70
python-dotenv==1.0.1